Add support for rendering complex scripts
Since Pango does not (yet) have a API to use a cairo_font_face_t for
rendering, HarfBuzz is used directly instead to do the actual text
layout, but Pango is still (ab)use to do text itemization.

The thumbnailer still does not do complex text, but I don’t think it
actually need to, so it was left alone.

https://bugzilla.gnome.org/show_bug.cgi?id=346942