Don’t select non-Unicode font charmaps
If the font does not contains all characters checked for, it may end up
with a non-Unicode charmap selected, since the code in
check_font_contain_text() never resets the original charmap. This will
break text layout with HarfBuzz and will give us empty boxes.

I fail to see the logic behind this code, we are only interested in the
Unicode charmap, since the others are pretty useless for us, especially
when there is a Unicode charmap in the font. So just let
FT_Select_Charmap() find the best Unicode charmap in the font for us.