font-model: Create cairo surfaces in the main thread
Commit b42c741768b20fb692631779eb1ec07fffaee0c1 switched from using
pixbufs to cairo surfaces so that we can scale them properly for HiDpi
displays but kept the cairo surface instantiations in the IO scheduler
job thread which means that, on X, we're using Xlib from multiple
threads.

Avoid this by only loading the pixbufs in the IO thread and then
creating the cairo surfaces on the main thread.

https://bugzilla.gnome.org/show_bug.cgi?id=720204