font-view: use GApplication commandline handling
Use GApplication's internal handling of commandline arguments.  This
means that --gapplication-service (which we depend on for DBus
activation) will be properly supported.

The only reason this was working before is because we were not actually
being D-Bus activated due to having a '-' in our app id.  With the fix
of bug 764754 this is no longer the case, and gnome-font-viewer will be
broken.

https://bugzilla.gnome.org/show_bug.cgi?id=765518