model: listen to fontconfig timestamp notifications
Instead of watching all the font directories manually, connect to the
fontconfig timestamp notification.
This also fixes a race condition where we were trying to recreate the
configuration of fontconfig before font files were completely written,
which caused installed fonts not to get picked up by fontconfig under
certain circumstances.

Note that this relies on an agent being present in the session updating
the Fontconfig/Timestamp XSETTING. In GNOME, gnome-settings-daemon takes
care of that.