font-view: account for font files with same basename
We make sure the font we're trying to install doesn't already
exist on the system by matching its metadata, but we copy the font file
keeping its original basename.

In case a font file with the same basename already exists in the user
directory, but for another font than the one we're trying to install,
we'll thus fail the installation, because the file already exists.

Account for that possibility by adding a number as a prefix to the
font file name, when that happens.

https://gitlab.gnome.org/GNOME/gnome-font-viewer/issues/10
https://gitlab.gnome.org/GNOME/gnome-font-viewer/issues/11